chore: initial config refactor
