+++ package.json
@@ -38,15 +38,15 @@
     "@ladjs/bull": "^1.0.5",
     "@ladjs/env": "^1.0.0",
     "@ladjs/graceful": "^1.0.0",
-    "@ladjs/i18n": "^3.0.10",
+    "@ladjs/i18n": "^3.0.11",
     "@ladjs/mongoose": "^2.0.1",
     "@ladjs/mongoose-error-messages": "^1.0.0",
     "@ladjs/passport": "^2.0.0",
     "@ladjs/policies": "^3.1.0",
     "@ladjs/proxy": "^1.0.3",
     "@ladjs/redis": "^1.0.3",
-    "@ladjs/shared-config": "^2.0.11",
-    "@ladjs/web": "^2.0.0",
+    "@ladjs/shared-config": "^2.0.13",
+    "@ladjs/web": "^2.0.3",
     "@primer/css": "^14.3.0",
     "@sidoshi/random-string": "^1.0.0",
     "@tkrotoff/bootstrap-floating-label": "^0.7.0",
@@ -84,7 +84,7 @@
     "koa-views-render": "^0.0.1",
     "lodash": "^4.17.15",
     "mandarin": "^2.0.5",
-    "manifest-rev": "^1.0.2",
+    "manifest-rev": "^1.0.3",
     "markdown-it": "^11.0.0",
     "markdown-it-emoji": "^1.4.0",
     "markdown-it-github-headings": "^2.0.0",
